# Conflicts:
#	app/src/main/java/com/hyeong/handsomego/applicationController/NetworkService.kt
#	app/src/main/java/com/hyeong/handsomego/home/HomeTabFragment.kt
#	app/src/main/res/layout/fragment_home_tab_rank.xml

+++ app/src/main/java/com/hyeong/handsomego/get/GetMoreReviewResponseData.kt
@@ -4,7 +4,7 @@ data class GetMoreReviewResponseData (
         var writer_name: String,
         var comment_id: Int,
         var writer_id: Int,
-        var comment_star: Int,
+        var comment_star: Float,
         var comment_date: String,
         var comment_comment: String,
         var place_id: Int,
